\name{sdef-package}
\alias{sdef-package}
\alias{sdef}
\docType{package}
\title{
\packageTitle{sdef}
}
\description{
\packageDescription{sdef}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{sdef}
\packageIndices{sdef}

}
\author{
\packageAuthor{sdef}

Maintainer: \packageMaintainer{sdef}
}
