\name{sdef-package}
\Rdversion{1.1}
\alias{sdef-package}
\alias{sdef}
\docType{package}
\title{Synthesizing List of Differentially Expressed Features}
\description{
The package performs two tests to evaluate if the experiments are associated and returns a list of interesting feautres common to all the experiments.
}
\details{
\tabular{ll}{
Package: \tab sdef\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2009-08-17\cr
License: \tab GPL-2\cr
Lazyload: \tab yes\cr
}
}
\author{
Marta Blangiardo, Alberto Cassese

Maintainer: Marta Blangiardo <m.blangiardo@imperial.ac.uk>, Alberto Cassese <fratucster@gmail.com>

}
