\name{sdef-package}
\Rdversion{1.2}
\alias{sdef-package}
\alias{sdef}
\docType{package}
\title{Synthesizing List of Differentially Expressed Features}
\description{
The package performs two tests to evaluate if the experiments are associated and returns a list of interesting features common to all the experiments.
}
\details{
\tabular{ll}{
Package: \tab sdef\cr
Type: \tab Package\cr
Version: \tab 1.2\cr
Date: \tab 2009-09-04\cr
License: \tab GPL-2\cr
Lazyload: \tab yes\cr
}
}
\author{
Alberto Cassese, Marta Blangiardo

Maintainer: Alberto Cassese <fratucster@gmail.com>

}
